requirements.txt
asyncio==3.4.3
asyncpg==0.20.1
gino==1.0.1
pkg-resources==0.0.0
python-dotenv==0.14.0
SQLAlchemy==1.3.18